update android project using react native upgrade
